drop.tip: Fix bug which causes a corrupt phylo object when subtree = T
The behaviour is reproducible on previous versions with:

    tree = read.tree(text = "(A:1,(B:1,(C:1,(D:1,E:1):1):1):1):1;")
    print(tree)
    print(drop.tip(tree, "D", subtree = T))
    print(drop.tip(tree, "D", subtree = T)$edge)

which outputs:

    Phylogenetic tree with 5 tips and 4 internal nodes.

    Tip labels:
    [1] "A" "B" "C" "D" "E"

    Rooted; includes branch lengths.

    Phylogenetic tree with 1 tips and 8 internal nodes.

    Tip labels:
    [1] "[_tips]"

    Rooted; includes branch lengths.
         [,1] [,2]
    [1,]    0    1
    [2,]    0    3
    [3,]    3    4
    [4,]    3    5
    [5,]    5    6
    [6,]    5    7
    [7,]    7    8
    [8,]    7    9

Operations on the resulting phylo object will result in errors due to
tip/edge mismatches.

There are two small bugs contributing to the behaviour described above,
which are triggered when subtree = T has caused the all the "dropped"
tips to be kept and merely re-labeled as "[1_tip]".

The first bug is that all tip labels are removed, because the vector of
actually dropped tips is empty (due to tip <- tip[-i] on line 163) and
subsetting a vector with [-integer(0)] returns an empty vector instead
of the original vector:

    > c(1,2,3)[-2]
    [1] 1 3
    > c(1,2,3)[-integer(0)]
    numeric(0)

The fix is guarding the subset operation with a length check.

The second bug is the incorrect assumption that there will always be
some nodes converted to tips, which causes the string "[_tips]", seen
above, to be added to the emptied tip label vector.  The fix is guarding
the paste() with a length check on node to tip conversions.